+++ eos_invoices/__init__.py
@@ -1,4 +1,4 @@
 """Outstanding payments of a Corporation across Alliance Auth apps."""
 
-__version__ = "0.0.4"
+__version__ = "0.0.5"
 VERSION = __version__